Move components from ui-component to admin-panel
