+++ packages/admin-panel/src/dataTables/components/PreviewFilters/filters/DatePicker.js
@@ -6,7 +6,7 @@
 import PropTypes from 'prop-types';
 import React, { useEffect } from 'react';
 
-import { DatePicker as BaseDatePicker } from '../../../Inputs';
+import { DatePicker as BaseDatePicker } from '@tupaia/ui-components';
 import { ParameterType } from '../../editing';
 
 export const DatePicker = ({ name, defaultValue, hasDefaultValue, inputFilterValue, onChange }) => {
