+++ packages/ui-components/src/components/DataTable/index.js
@@ -5,5 +5,3 @@
 
 export * from './DataTable';
 export * from './useDataTableExport';
-export { ParameterList, ParameterItem } from './editing';
-export { PreviewFilters } from './PreviewFilters';
